backend/requirements.txt
Django>=3.2.6,<3.3.0
djangorestframework>=3.12.4,<3.13.0
django-cors-headers>=3.8.0,<3.9.0
flake8>=3.9.2,<3.10.0 
gunicorn>=20.1.0,<20.2.0
